+++ lib/msal-browser/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @azure/msal-browser
 
-This log was last generated on Mon, 07 Aug 2023 18:15:15 GMT and should not be manually modified.
+This log was last generated on Fri, 18 Aug 2023 18:40:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 3.0.2
+
+Fri, 18 Aug 2023 18:40:03 GMT
+
+### Patches
+
+- Fix input parameter type for addEventCallback (thomas.norling@microsoft.com)
+- Extend hydrateCache request type to include ssoSilentRequest, PopupRequest and RedirectRequest #6329 (thomas.norling@microsoft.com)
+- Bump @azure/msal-common to v14.0.2
+
+## 3.0.1
+
+Fri, 11 Aug 2023 19:00:44 GMT
+
+### Patches
+
+- Update dist settings for packages #6322 (hemoral@microsoft.com)
+- Add `tokenBodyParameters` to silent and interaction request types #6325 (kshabelko@microsoft.com)
+- Bump @azure/msal-common to v14.0.1
+
+### Changes
+
+- Pick up broker extra query params #6286 (kshabelko@microsoft.com)
 
 ## 3.0.0